minor correction to regex
a negative lookbehind i added to stop incorrect matching of some malformed numeric literals, not matching if following a "." doesn't match when following a range ".." operator, which *is* a valid syntax. this changes the lookbehind to check two characters and make sure the char before the one directly preceding the literal is not a "." so that 0..10 works as expected. this may cause an issue with custom operators that end with a "." and could be tweaked further, this is just a quick fix.